infrastructure/services/profile-service/backend/app/utils/redis_cache.py
import json
import logging
import os
from typing import Any, Optional

import redis

logger = logging.getLogger(__name__)

REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379")
TTL_MEMBER_GET = int(os.getenv("REDIS_TTL_MEMBER_GET", 60))
TTL_MEMBER_SEARCH = int(os.getenv("REDIS_TTL_MEMBER_SEARCH", 30))

_client: Optional[redis.Redis] = None


def get_redis() -> Optional[redis.Redis]:
    global _client
    if _client is None:
        try:
            _client = redis.from_url(REDIS_URL, decode_responses=True, socket_connect_timeout=2)
            _client.ping()
        except Exception as e:
            logger.warning("Redis unavailable, caching disabled: %s", e)
            _client = None
    return _client


def cache_get(key: str) -> Optional[Any]:
    r = get_redis()
    if r is None:
        return None
    try:
        val = r.get(key)
        return json.loads(val) if val else None
    except Exception as e:
        logger.warning("Redis GET error: %s", e)
        return None


def cache_set(key: str, value: Any, ttl: int) -> None:
    r = get_redis()
    if r is None:
        return
    try:
        r.setex(key, ttl, json.dumps(value, default=str))
    except Exception as e:
        logger.warning("Redis SET error: %s", e)


def cache_delete(key: str) -> None:
    r = get_redis()
    if r is None:
        return
    try:
        r.delete(key)
    except Exception as e:
        logger.warning("Redis DEL error: %s", e)


def cache_delete_pattern(pattern: str) -> None:
    r = get_redis()
    if r is None:
        return
    try:
        keys = r.keys(pattern)
        if keys:
            r.delete(*keys)
    except Exception as e:
        logger.warning("Redis DEL pattern error: %s", e)
